requirements.txt
<< 웹 서버 >>
flask==1.1.2
Flask-SocketIO==5.0.1
tesorflow==2.4.1
pandas==1.1.3
jpype1=1.2.0

<< 입력 데이터 전처리 >>
hgtk==0.1.3
soynlp==0.0493
konlpy==0.5.2
PIL==8.0.1
urllib==1.25.11
numpy==1.19.2


